// src/types/generated/matrixEnjinV1012.ts
export interface CounterOffer {
    sellerPrice: bigint
    buyerPrice?: (bigint | undefined)
    deposit: Deposit
}

export interface Deposit {
    depositor: AccountId32
    amount: bigint
}

export interface Listing {
    creator: AccountId32
    makeAssetId: AssetId
    takeAssetId: AssetId
    amount: bigint
    price: bigint
    minReceived: bigint
    feeSide: FeeSide
    creationBlock: number
    deposit: Deposit
    salt: Bytes
    data: ListingData
    state: ListingState
}

export type ListingState = ListingState_Auction | ListingState_FixedPrice | ListingState_Offer

export interface ListingState_Auction {
    __kind: 'Auction'
    value: AuctionState
}

export interface ListingState_FixedPrice {
    __kind: 'FixedPrice'
    amountFilled: bigint
}

export interface ListingState_Offer {
    __kind: 'Offer'
    value: OfferState
}

export interface OfferState {
    counterOfferCount: number
}

export interface AuctionState {
    highBid?: (Bid | undefined)
}

export interface Bid {
    bidder: AccountId32
    price: bigint
}

export type ListingData = ListingData_Auction | ListingData_FixedPrice | ListingData_Offer

export interface ListingData_Auction {
    __kind: 'Auction'
    value: AuctionData
}

export interface ListingData_FixedPrice {
    __kind: 'FixedPrice'
}

export interface ListingData_Offer {
    __kind: 'Offer'
    value: OfferData
}

export interface OfferData {
    expiration?: (number | undefined)
}

export interface AuctionData {
    startBlock: number
    endBlock: number
}

export type FeeSide = FeeSide_Make | FeeSide_NoFee | FeeSide_Take

export interface FeeSide_Make {
    __kind: 'Make'
}

export interface FeeSide_NoFee {
    __kind: 'NoFee'
}

export interface FeeSide_Take {
    __kind: 'Take'
}
// ...
export interface AssetId {
    collectionId: bigint
    tokenId: bigint
}
// ...
export type CounterOfferResponse = CounterOfferResponse_Accept | CounterOfferResponse_Counter | CounterOfferResponse_Reject

export interface CounterOfferResponse_Accept {
    __kind: 'Accept'
}

export interface CounterOfferResponse_Counter {
    __kind: 'Counter'
    value: bigint
}

export interface CounterOfferResponse_Reject {
    __kind: 'Reject'
}
// ...
export type AccountId32 = Bytes
// ...
export type H256 = Bytes
// ...
export const ListingData: sts.Type<ListingData> = sts.closedEnum(() => {
    return  {
        Auction: AuctionData,
        FixedPrice: sts.unit(),
        Offer: OfferData,
    }
})

export const OfferData: sts.Type<OfferData> = sts.struct(() => {
    return  {
        expiration: sts.option(() => sts.number()),
    }
})

export const AuctionData: sts.Type<AuctionData> = sts.struct(() => {
    return  {
        startBlock: sts.number(),
        endBlock: sts.number(),
    }
})
// ...
export const AssetId: sts.Type<AssetId> = sts.struct(() => {
    return  {
        collectionId: sts.bigint(),
        tokenId: sts.bigint(),
    }
})
// ...
export const CounterOfferResponse: sts.Type<CounterOfferResponse> = sts.closedEnum(() => {
    return  {
        Accept: sts.unit(),
        Counter: sts.bigint(),
        Reject: sts.unit(),
    }
})

export const CounterOffer: sts.Type<CounterOffer> = sts.struct(() => {
    return  {
        sellerPrice: sts.bigint(),
        buyerPrice: sts.option(() => sts.bigint()),
        deposit: Deposit,
    }
})

export const Deposit: sts.Type<Deposit> = sts.struct(() => {
    return  {
        depositor: AccountId32,
        amount: sts.bigint(),
    }
})

export const Listing: sts.Type<Listing> = sts.struct(() => {
    return  {
        creator: AccountId32,
        makeAssetId: AssetId,
        takeAssetId: AssetId,
        amount: sts.bigint(),
        price: sts.bigint(),
        minReceived: sts.bigint(),
        feeSide: FeeSide,
        creationBlock: sts.number(),
        deposit: Deposit,
        salt: sts.bytes(),
        data: ListingData,
        state: ListingState,
    }
})

export const ListingState: sts.Type<ListingState> = sts.closedEnum(() => {
    return  {
        Auction: AuctionState,
        FixedPrice: sts.enumStruct({
            amountFilled: sts.bigint(),
        }),
        Offer: OfferState,
    }
})

export const OfferState: sts.Type<OfferState> = sts.struct(() => {
    return  {
        counterOfferCount: sts.number(),
    }
})

export const AuctionState: sts.Type<AuctionState> = sts.struct(() => {
    return  {
        highBid: sts.option(() => Bid),
    }
})

export const FeeSide: sts.Type<FeeSide> = sts.closedEnum(() => {
    return  {
        Make: sts.unit(),
        NoFee: sts.unit(),
        Take: sts.unit(),
    }
})
// ...
export const AccountId32 = sts.bytes()

export const H256 = sts.bytes()